Use seperate class to read nfo subtrees
