+++ MediaBrowser.XbmcMetadata/Parsers/BaseNfoParser.cs
@@ -588,23 +588,9 @@ namespace MediaBrowser.XbmcMetadata.Parsers
                         break;
                     }
 
-                // todo
                 case "fileinfo":
-                    {
-                        if (!reader.IsEmptyElement)
-                        {
-                            using (var subtree = reader.ReadSubtree())
-                            {
-                                FetchFromFileInfoNode(subtree, item);
-                            }
-                        }
-                        else
-                        {
-                            reader.Read();
-                        }
-
-                        break;
-                    }
+                    NfoSubtreeParsers.ReadFileinfoSubtree(reader, item);
+                    break;
 
                 // todo
                 case "uniqueid":
@@ -687,243 +673,6 @@ namespace MediaBrowser.XbmcMetadata.Parsers
                 default:
                     reader.ReadProviderIdFromNfo(item, _validProviderIds);
                     break;
-            }
-        }
-
-        private void FetchFromFileInfoNode(XmlReader reader, T item)
-        {
-            reader.MoveToContent();
-            reader.Read();
-
-            // Loop through each element
-            while (!reader.EOF && reader.ReadState == ReadState.Interactive)
-            {
-                if (reader.NodeType == XmlNodeType.Element)
-                {
-                    switch (reader.Name)
-                    {
-                        case "streamdetails":
-                            {
-                                if (reader.IsEmptyElement)
-                                {
-                                    reader.Read();
-                                    continue;
-                                }
-
-                                using (var subtree = reader.ReadSubtree())
-                                {
-                                    FetchFromStreamDetailsNode(subtree, item);
-                                }
-
-                                break;
-                            }
-
-                        default:
-                            reader.Skip();
-                            break;
-                    }
-                }
-                else
-                {
-                    reader.Read();
-                }
-            }
-        }
-
-        private void FetchFromStreamDetailsNode(XmlReader reader, T item)
-        {
-            reader.MoveToContent();
-            reader.Read();
-
-            // Loop through each element
-            while (!reader.EOF && reader.ReadState == ReadState.Interactive)
-            {
-                if (reader.NodeType == XmlNodeType.Element)
-                {
-                    switch (reader.Name)
-                    {
-                        case "video":
-                            {
-                                if (reader.IsEmptyElement)
-                                {
-                                    reader.Read();
-                                    continue;
-                                }
-
-                                using (var subtree = reader.ReadSubtree())
-                                {
-                                    FetchFromVideoNode(subtree, item);
-                                }
-
-                                break;
-                            }
-
-                        case "subtitle":
-                            {
-                                if (reader.IsEmptyElement)
-                                {
-                                    reader.Read();
-                                    continue;
-                                }
-
-                                using (var subtree = reader.ReadSubtree())
-                                {
-                                    FetchFromSubtitleNode(subtree, item);
-                                }
-
-                                break;
-                            }
-
-                        default:
-                            reader.Skip();
-                            break;
-                    }
-                }
-                else
-                {
-                    reader.Read();
-                }
-            }
-        }
-
-        private void FetchFromVideoNode(XmlReader reader, T item)
-        {
-            reader.MoveToContent();
-            reader.Read();
-
-            // Loop through each element
-            while (!reader.EOF && reader.ReadState == ReadState.Interactive)
-            {
-                if (reader.NodeType == XmlNodeType.Element)
-                {
-                    switch (reader.Name)
-                    {
-                        case "format3d":
-                            {
-                                var val = reader.ReadElementContentAsString();
-
-                                var video = item as Video;
-
-                                if (video != null)
-                                {
-                                    if (string.Equals("HSBS", val, StringComparison.OrdinalIgnoreCase))
-                                    {
-                                        video.Video3DFormat = Video3DFormat.HalfSideBySide;
-                                    }
-                                    else if (string.Equals("HTAB", val, StringComparison.OrdinalIgnoreCase))
-                                    {
-                                        video.Video3DFormat = Video3DFormat.HalfTopAndBottom;
-                                    }
-                                    else if (string.Equals("FTAB", val, StringComparison.OrdinalIgnoreCase))
-                                    {
-                                        video.Video3DFormat = Video3DFormat.FullTopAndBottom;
-                                    }
-                                    else if (string.Equals("FSBS", val, StringComparison.OrdinalIgnoreCase))
-                                    {
-                                        video.Video3DFormat = Video3DFormat.FullSideBySide;
-                                    }
-                                    else if (string.Equals("MVC", val, StringComparison.OrdinalIgnoreCase))
-                                    {
-                                        video.Video3DFormat = Video3DFormat.MVC;
-                                    }
-                                }
-
-                                break;
-                            }
-
-                        case "aspect":
-                            {
-                                var val = reader.ReadElementContentAsString();
-
-                                if (item is Video video)
-                                {
-                                    video.AspectRatio = val;
-                                }
-
-                                break;
-                            }
-
-                        case "width":
-                            {
-                                var val = reader.ReadElementContentAsInt();
-
-                                if (item is Video video)
-                                {
-                                    video.Width = val;
-                                }
-
-                                break;
-                            }
-
-                        case "height":
-                            {
-                                var val = reader.ReadElementContentAsInt();
-
-                                if (item is Video video)
-                                {
-                                    video.Height = val;
-                                }
-
-                                break;
-                            }
-
-                        case "durationinseconds":
-                            {
-                                var val = reader.ReadElementContentAsInt();
-
-                                if (item is Video video)
-                                {
-                                    video.RunTimeTicks = new TimeSpan(0, 0, val).Ticks;
-                                }
-
-                                break;
-                            }
-
-                        default:
-                            reader.Skip();
-                            break;
-                    }
-                }
-                else
-                {
-                    reader.Read();
-                }
-            }
-        }
-
-        private void FetchFromSubtitleNode(XmlReader reader, T item)
-        {
-            reader.MoveToContent();
-            reader.Read();
-
-            // Loop through each element
-            while (!reader.EOF && reader.ReadState == ReadState.Interactive)
-            {
-                if (reader.NodeType == XmlNodeType.Element)
-                {
-                    switch (reader.Name)
-                    {
-                        case "language":
-                            {
-                                _ = reader.ReadElementContentAsString();
-
-                                if (item is Video video)
-                                {
-                                    video.HasSubtitles = true;
-                                }
-
-                                break;
-                            }
-
-                        default:
-                            reader.Skip();
-                            break;
-                    }
-                }
-                else
-                {
-                    reader.Read();
-                }
             }
         }
 
@@ -1038,6 +787,7 @@ namespace MediaBrowser.XbmcMetadata.Parsers
             };
         }
 
+        // todo remove
         internal XmlReaderSettings GetXmlReaderSettings()
             => new XmlReaderSettings()
             {
